refactor: move errors out of the api package to fix its root cause
The circular-import fix in the previous commit worked but was the wrong
shape for it: ~19 local imports scattered across 6 files is not what
PEP 8's "avoid circular imports" exception is for, and it hid the actual
problem instead of removing it.

The real issue was structural: api/errors/ was a subpackage of api, so
importing any error type required Python to run api/__init__.py first -
which does `from learninghouse.api import auth, brain, sensor`, pulling in
models/services that import back from api.errors. Whichever module a test
or script imported first decided whether that succeeded.

api/errors/__init__.py never depended on anything in the api package to
begin with - only on core and models - so it was never entangled with the
api package, just living inside it. Moved it to a top-level `errors`
module instead, and reverted every local import from the previous commit
back to the module level: models, services and api can all depend on
errors now without any of them needing the other two loaded first.
