+++ core/learninghouse/api/ui.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from fastapi import APIRouter, Request
 from fastapi.responses import FileResponse, PlainTextResponse, RedirectResponse
 
-from learninghouse.api.errors import LearningHouseSecurityException
+from learninghouse.errors import LearningHouseSecurityException
 
 router = APIRouter(include_in_schema=False)
