+++ core/learninghouse/errors/sensor.py
@@ -2,7 +2,7 @@ from typing import Any, Dict
 
 from fastapi import status
 
-from learninghouse.api.errors import LearningHouseException
+from learninghouse.errors import LearningHouseException
 from learninghouse.models import LearningHouseErrorMessage
 
 MIMETYPE_JSON = "application/json"
