+++ core/learninghouse/services/brain.py
@@ -10,6 +10,14 @@ from sklearn.metrics import accuracy_score
 
 from learninghouse.core.logger import logger
 from learninghouse.core.settings import service_settings
+from learninghouse.errors.brain import (
+    BrainBadRequest,
+    BrainExists,
+    BrainNoConfiguration,
+    BrainNotActual,
+    BrainNotEnoughData,
+    BrainNotTrained,
+)
 from learninghouse.models.brain import (
     Brain,
     BrainConfiguration,
@@ -32,8 +40,6 @@ class BrainService:
 
     @classmethod
     def list_all(cls) -> BrainInfos:
-        from learninghouse.api.errors.brain import BrainNoConfiguration
-
         brains: Dict[str, BrainInfo] = {}
         for directory in listdir(service_settings().brains_directory):
             try:
@@ -45,8 +51,6 @@ class BrainService:
 
     @staticmethod
     def get_info(name: str) -> BrainInfo:
-        from learninghouse.api.errors.brain import BrainNoConfiguration, BrainNotTrained
-
         info: Optional[BrainInfo] = None
         if Brain.is_trained(name):
             try:
@@ -75,8 +79,6 @@ class BrainService:
         dependent_value: Optional[Any] = None,
         sensors_data: Optional[Dict[str, Any]] = None,
     ) -> BrainInfo:
-        from learninghouse.api.errors.brain import BrainBadRequest, BrainNotEnoughData
-
         filename = Brain.sanitize_filename(name, BrainFileType.TRAINING_DATA_FILE)
 
         trainings_data: Optional[Dict[str, Any]] = sensors_data
@@ -110,11 +112,6 @@ class BrainService:
 
     @staticmethod
     def train(name: str, data: pd.DataFrame) -> BrainInfo:
-        from learninghouse.api.errors.brain import (
-            BrainNoConfiguration,
-            BrainNotEnoughData,
-        )
-
         try:
             brain = Brain(name)
 
@@ -162,8 +159,6 @@ class BrainService:
 
     @classmethod
     def prediction(cls, name: str, request_data: Dict[str, Any]):
-        from learninghouse.api.errors.brain import BrainNotActual, BrainNotTrained
-
         try:
             brain = cls.load_brain(name)
             if not brain.actual_versions:
@@ -212,8 +207,6 @@ class BrainService:
 class BrainConfigurationService:
     @staticmethod
     def get(name: str) -> BrainConfiguration:
-        from learninghouse.api.errors.brain import BrainNoConfiguration
-
         try:
             return BrainConfiguration.from_json_file(name)
         except FileNotFoundError as exc:
@@ -221,8 +214,6 @@ class BrainConfigurationService:
 
     @staticmethod
     def create(configuration: BrainConfiguration) -> BrainConfiguration:
-        from learninghouse.api.errors.brain import BrainExists
-
         if BrainConfiguration.json_config_file_exists(configuration.name):
             raise BrainExists(configuration.name)
 
@@ -232,8 +223,6 @@ class BrainConfigurationService:
 
     @staticmethod
     def update(name: str, configuration: BrainConfiguration) -> BrainConfiguration:
-        from learninghouse.api.errors.brain import BrainNoConfiguration
-
         if not BrainConfiguration.json_config_file_exists(name):
             raise BrainNoConfiguration(name)
 
@@ -243,8 +232,6 @@ class BrainConfigurationService:
 
     @staticmethod
     def delete(name: str) -> BrainDeleteResult:
-        from learninghouse.api.errors.brain import BrainNoConfiguration
-
         brainpath = Brain.sanitize_directory(name)
 
         if not path.exists(brainpath):
