+++ core/learninghouse/services/sensor.py
@@ -1,3 +1,4 @@
+from learninghouse.errors.sensor import NoSensor, SensorExists
 from learninghouse.models.sensor import (
     Sensor,
     SensorDeleteResult,
@@ -13,8 +14,6 @@ class SensorConfigurationService:
 
     @staticmethod
     def get(name: str) -> Sensor:
-        from learninghouse.api.errors.sensor import NoSensor
-
         sensors = Sensors.load_config()
 
         for sensor in sensors.root:
@@ -25,8 +24,6 @@ class SensorConfigurationService:
 
     @staticmethod
     def create(name: str, typed: SensorType) -> Sensor:
-        from learninghouse.api.errors.sensor import SensorExists
-
         sensors = Sensors.load_config()
 
         for sensor in sensors.root:
@@ -43,8 +40,6 @@ class SensorConfigurationService:
     def update(
         name: str, typed: SensorType, cycles: int, calc_sun_position: bool
     ) -> Sensor:
-        from learninghouse.api.errors.sensor import NoSensor
-
         sensors = Sensors.load_config()
         for sensor in sensors.root:
             if sensor.name == name:
@@ -58,8 +53,6 @@ class SensorConfigurationService:
 
     @staticmethod
     def delete(name: str) -> SensorDeleteResult:
-        from learninghouse.api.errors.sensor import NoSensor
-
         sensors = Sensors.load_config()
 
         for sensor in sensors.root:
